Refactor model.py: split it into two modules
- model.py: generic models that can be used for timelag/duration/acoustic
models
- acoustic_models.py: acoustic models
